meta-isar/recipes-app: change SRC_URIs of hello rebuild example
A following change will require the PV to be the exact match of the
upstream PV. So "2.10" will not work because upsteam is at "2.10-2", and
similar for "2.9".
We do not want the examples to rebuild an exact version, but rather the
most recent from the suite. The "=" was in there mainly to document the
choosing feature, which is now done by the comment.

Signed-off-by: Henning Schild <henning.schild@siemens.com>