dpkg-base: ignore binary packages for "apt-get source"
A buildchroot might already know the new/rebuild version of an upstream
package as a binary coming from isar-apt. Now apt-get source will try
to fetch that one instead of the upstream version that we rebuild to
produce the new one.

Not that ignoring binaries also means we loose the vague version
matching, inform users about that.

This problem can be seen in partial rebuild scenarios and the switch
solves it, at least until we do have sources in isar-apt. Having sources
there, the best idea would probably be to choose the exact upstream
version to rebuild, to not find our own in an incremental build.

Signed-off-by: Henning Schild <henning.schild@siemens.com>