recipes-kernel/linux: make KERNEL_DEFCONFIG support in-tree defconfigs
The recipe expected the specified KERNEL_DEFCONFIG to be provided as a
SRC_URI while the kernel already provides many useful defconfigs. The
recipe will now check if the specified configuration can be found in
WORKDIR. If it is found, it will then be used via "make olddefconfig"
as before. It will otherwise assume that it is an in-tree defconfig
and attempt a "make foobar_defconfig" (where foobar_defconfig is the
user-provided KERNEL_DEFCONFIG setting). To support this change, a
KERNEL_CONFIG_TARGET variable is passed to the build-kernel.sh script
via the environment.

Signed-off-by: Cedric Hombourger <Cedric_Hombourger@mentor.com>